requirements.txt
Django==4.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
psycopg2-binary==2.9.6
drf-yasg==1.21.5
gunicorn==21.0.1  # Para execução em produção com WSGI
python-decouple==3.8  # Para gerenciar variáveis de ambiente
